Fix tests after rebase with master
